src/generate_summary/requirements.txt
line-bot-sdk
boto3
langchain
bs4
openai
langchain-community
langchain-core
langchain-openai
youtube-transcript-api